split into prayerpointservice and complexPrayerOperations
